Split the reordering tests into two files (for widgets and dashboards, respectively) and added removing the boards after the test is done to dashboards' tests.
